Fix : 結果画面を静的ファイルと動的ファイルに分割
